Combined Modules. Added translators
