--- hexagonal-arch-demo/core/src/main/java/com/baeldung/demo/model/Customer.java
@@ -1,18 +1,0 @@
-package com.baeldung.demo.model;
-
-public class Customer {
-    public String firstName;
-    public String lastName;
-    public String zipCode;
-    public String emailId;
-    public String ssn;
-    // Standard getters and setters
-    public boolean isCreditCheckRequired() {
-        // Business logic that determines based on SSN
-        return false;
-    }
-    public boolean isFreeShippingEligible() {
-        // Business logic that determines free shipping based on zip code
-        return false;
-    }
-}
