--- hexagonal-arch-demo/jpa-adapter/src/main/java/com/baeldung/demo/customer/jpaadapter/repository/CustomerRepository.java
@@ -1,9 +1,0 @@
-package com.baeldung.demo.customer.jpaadapter.repository;
-
-import com.baeldung.demo.model.Customer;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface CustomerRepository extends JpaRepository<Customer, String> {
-}
