--- hexagonal-arch-demo/port/pom.xml
@@ -1,40 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project xmlns="http://maven.apache.org/POM/4.0.0"
-         xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-    <modelVersion>4.0.0</modelVersion>
-
-    <groupId>com.baeldung.demo.customer</groupId>
-    <artifactId>port</artifactId>
-    <version>1.0-SNAPSHOT</version>
-    <description>Primary Ports for the application</description>
-
-    <dependencies>
-        <dependency>
-            <groupId>com.baeldung.demo.customer</groupId>
-            <artifactId>core</artifactId>
-            <version>1.0-SNAPSHOT</version>
-        </dependency>
-        <!-- Secondary Adapter needed by the application-->
-        <dependency>
-            <groupId>com.baeldung.demo.customer</groupId>
-            <artifactId>jpa-adapter</artifactId>
-            <version>0.0.1-SNAPSHOT</version>
-        </dependency>
-        <!-- Secondary Port needed by the application-->
-        <dependency>
-            <groupId>com.baeldung.demo.customer</groupId>
-            <artifactId>secondary-port</artifactId>
-            <version>1.0-SNAPSHOT</version>
-        </dependency>
-        <!-- https://mvnrepository.com/artifact/org.springframework/spring-beans -->
-        <!--<dependency>-->
-            <!--<groupId>org.springframework</groupId>-->
-            <!--<artifactId>spring-beans</artifactId>-->
-            <!--<version>5.1.5.RELEASE</version>-->
-        <!--</dependency>-->
-
-    </dependencies>
-
-
-</project>
