--- hexagonal-arch-demo/port/src/main/java/com/baeldung/demo/port/primary/CustomerService.java
@@ -1,10 +1,0 @@
-package com.baeldung.demo.port.primary;
-
-import com.baeldung.demo.model.Customer;
-
-public interface CustomerService {
-    Customer createCustomer(Customer customer);
-    Customer updateCustomerAddress(Customer customer);
-    boolean isFreeShippingEligible(String customerId);
-    Customer getCustomer(String customerId);
-}
