--- hexagonal-arch-demo/rest-adapter/src/main/java/com/baeldung/demo/customer/restadapter/RestAdapterApplication.java
@@ -1,13 +1,0 @@
-package com.baeldung.demo.customer.restadapter;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class RestAdapterApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(RestAdapterApplication.class, args);
-	}
-
-}
