--- hexagonal-arch-demo/rest-adapter/src/main/resources/application.properties
@@ -1,1 +1,0 @@
-
