--- hexagonal-arch-demo/secondary-port/pom.xml
@@ -1,22 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project xmlns="http://maven.apache.org/POM/4.0.0"
-         xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-    <modelVersion>4.0.0</modelVersion>
-
-    <groupId>com.baeldung.demo.customer</groupId>
-    <artifactId>secondary-port</artifactId>
-    <version>1.0-SNAPSHOT</version>
-    <description>Secondary Ports for the application</description>
-
-    <dependencies>
-        <dependency>
-            <groupId>com.baeldung.demo.customer</groupId>
-            <artifactId>core</artifactId>
-            <version>1.0-SNAPSHOT</version>
-        </dependency>
-
-    </dependencies>
-
-
-</project>
