--- hexagonal-arch-demo/secondary-port/src/main/java/com/baeldung/demo/port/secondary/CustomerCommunicationService.java
@@ -1,6 +1,0 @@
-package com.baeldung.demo.port.secondary;
-
-public interface CustomerCommunicationService {
-    void notifyCrm();
-    int sendWelcomeEmail(String emailId);
-}
