--- hexagonal-arch-demo/secondary-port/src/main/java/com/baeldung/demo/port/secondary/CustomerPersistenceService.java
@@ -1,9 +1,0 @@
-package com.baeldung.demo.port.secondary;
-
-import com.baeldung.demo.model.Customer;
-
-public interface CustomerPersistenceService {
-    Customer createCustomer(Customer customer);
-    Customer retrieveCustomer(String customerId);
-    Customer updateCustomer(Customer customer);
-}
