--- hexagonal-arch-demo/smtp-adapter/pom.xml
@@ -1,13 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project xmlns="http://maven.apache.org/POM/4.0.0"
-         xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-    <modelVersion>4.0.0</modelVersion>
-
-    <groupId>com.baeldung.demo.customer</groupId>
-    <artifactId>smtp-adapter</artifactId>
-    <version>1.0-SNAPSHOT</version>
-    <description>One of the 3 secondary adapter/interface for the application</description>
-
-
-</project>
