+++ hexagonal-arch-demo/src/main/java/com/baeldung/hex/demo/adapter/primary/rest/resource/CustomerResource.java
@@ -1,0 +1,9 @@
+package com.baeldung.hex.demo.adapter.primary.rest.resource;
+
+public class CustomerResource {
+    public String fName;
+    public String lName;
+    public String zipCd;
+    public String emailAddress;
+    public String ssn;
+}
