+++ hexagonal-arch-demo/src/main/java/com/baeldung/hex/demo/port/primary/CustomerService.java
@@ -1,0 +1,11 @@
+package com.baeldung.hex.demo.port.primary;
+
+
+import com.baeldung.hex.demo.core.model.Customer;
+
+public interface CustomerService {
+    Customer createCustomer(Customer customer);
+    Customer updateCustomerAddress(Customer customer);
+    boolean isFreeShippingEligible(String customerId);
+    Customer getCustomer(String customerId);
+}
